+++ packages/app-core/src/index.ts
@@ -1,4 +1,20 @@
-import extendAuthenticationModel from './authenticationModel'
+import extendAuthenticationModel, {
+  initInternetAccounts,
+} from './authenticationModel'
+import {
+  addUndoKeyboardShortcuts,
+  initUndoModel,
+  undoMenuItems,
+} from './addUndo'
 import App from './App'
+import extendMenuModel from './topLevelMenuUtils'
 
-export { App, extendAuthenticationModel }
+export {
+  App,
+  addUndoKeyboardShortcuts,
+  initUndoModel,
+  initInternetAccounts,
+  undoMenuItems,
+  extendAuthenticationModel,
+  extendMenuModel,
+}
